Replace main.py with fastapi
